// src/types/data-requeried-sub.ts
import { OctatetosEntrada } from "./solution";

/* export interface dataForSubneteo {
    numRedes: number;
    oct1: number;
    octatetos_entrada: OctatetosEntrada;
} */

export interface DataRequireSubneteo {
    oct1: string;
    oct2: string;
    oct3: string;
    oct4: string;
    numRedes: string;
}